Refactored calculatePercentage function into utils
