+++ src/custom/state/claim/hooks/utils.ts
@@ -1,7 +1,9 @@
-import { CurrencyAmount, Token } from '@uniswap/sdk-core'
+import { Currency, CurrencyAmount, Percent, Token } from '@uniswap/sdk-core'
 
 import { SupportedChainId } from 'constants/chains'
 import { GNO, GpEther, USDC_BY_CHAIN, V_COW } from 'constants/tokens'
+import { ONE_HUNDRED_PERCENT, ZERO_PERCENT } from 'constants/misc'
+import { PERCENTAGE_PRECISION } from 'constants/index'
 
 import {
   CLAIMS_REPO,
@@ -15,6 +17,7 @@ import {
   VCowPrices,
 } from 'state/claim/hooks/index'
 
+import { formatSmart } from 'utils/format'
 /**
  * Helper function to check whether any claim is an investment option
  *
@@ -173,4 +176,20 @@ export function claimTypeToTokenAmount(type: ClaimType, chainId: SupportedChainI
     default:
       return undefined
   }
+}
+
+/**
+ * Helper function to calculate and return the percentage between 2 CurrencyAmount instances
+ */
+export function calculatePercentage<C1 extends Currency, C2 extends Currency>(
+  numerator: CurrencyAmount<C1>,
+  denominator: CurrencyAmount<C2>
+): string {
+  let percentage = denominator.equalTo(ZERO_PERCENT)
+    ? ZERO_PERCENT
+    : new Percent(numerator.quotient, denominator.quotient)
+  if (percentage.greaterThan(ONE_HUNDRED_PERCENT)) {
+    percentage = ONE_HUNDRED_PERCENT
+  }
+  return formatSmart(percentage, PERCENTAGE_PRECISION) || '0'
 }
